requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery==2.32.0
pandas==1.3.3
numpy==1.21.2
pyarrow